+++ Source/WebCore/webaudio/DelayNode.h
@@ -35,15 +35,15 @@ class AudioParam;
 
 class DelayNode : public AudioBasicProcessorNode {
 public:
-    static PassRefPtr<DelayNode> create(AudioContext* context, double sampleRate)
+    static PassRefPtr<DelayNode> create(AudioContext* context, float sampleRate)
     {
-        return adoptRef(new DelayNode(context, sampleRate));      
+        return adoptRef(new DelayNode(context, sampleRate));
     }
 
     AudioParam* delayTime();
 
 private:
-    DelayNode(AudioContext*, double sampleRate);
+    DelayNode(AudioContext*, float sampleRate);
 
     DelayProcessor* delayProcessor() { return static_cast<DelayProcessor*>(processor()); }
 };
